requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
Django==3.0.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
freeze==1.0.10
graphene==2.1.8
graphene-django==2.10.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
Pillow==7.1.2